requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
requests
behave
allure-behave
jsonschema
